refactor(ctld): reorganize CraneCtld sources into domain subdirectories
Split the flat src/CraneCtld/ layout into focused subdirectories:
- Account/: accounting type definitions (AccountDefs.h extracted from
  CtldPublicDefs.h) and AccountManager
- Accounting/: runtime accounting managers (AccountMetaContainer,
  LicenseManager renamed from LicensesManager)
- Database/: DB clients (DbClient, EmbeddedDbClient)

No logic changes; consumers updated to use direct includes.
