+++ src/CraneCtld/JobScheduler.cpp
@@ -21,12 +21,12 @@
 #include <absl/time/internal/cctz/src/time_zone_if.h>
 #include <google/protobuf/util/time_util.h>
 
-#include "AccountManager.h"
-#include "AccountMetaContainer.h"
+#include "Account/AccountManager.h"
+#include "Accounting/AccountMetaContainer.h"
+#include "Accounting/LicenseManager.h"
 #include "CranedMetaContainer.h"
 #include "CtldPublicDefs.h"
-#include "EmbeddedDbClient.h"
-#include "LicensesManager.h"
+#include "Database/EmbeddedDbClient.h"
 #include "Lua/LuaJobHandler.h"
 #include "RpcService/CranedKeeper.h"
 #include "crane/PluginClient.h"
@@ -835,7 +835,7 @@ void JobScheduler::PutRecoveredJobIntoRunningQueueLock_(
                                              job->AllocatedRes());
   }
   if (!job->licenses_count.empty())
-    g_licenses_manager->MallocLicenseWhenRecoverRunning(job->licenses_count);
+    g_license_manager->MallocLicenseWhenRecoverRunning(job->licenses_count);
 
   // The order of LockGuards matters.
   LockGuard running_guard(&m_running_job_map_mtx_);
@@ -1196,7 +1196,7 @@ void JobScheduler::ScheduleThread_() {
           }
 
           if (!job_in_scheduler->actual_licenses.empty()) {
-            if (!g_licenses_manager->MallocLicense(
+            if (!g_license_manager->MallocLicense(
                     job_in_scheduler->actual_licenses)) {
               job->pending_reason = "Licenses";
               continue;
@@ -1208,8 +1208,7 @@ void JobScheduler::ScheduleThread_() {
               !result) {
             // free licenses
             if (!job_in_scheduler->actual_licenses.empty()) {
-              g_licenses_manager->FreeLicense(
-                  job_in_scheduler->actual_licenses);
+              g_license_manager->FreeLicense(job_in_scheduler->actual_licenses);
             }
             job->pending_reason = result.error();
             continue;
@@ -1538,7 +1537,7 @@ void JobScheduler::ScheduleThread_() {
                                                    job->JobId());
           g_account_meta_container->FreeQosResource(*job);
           if (!job->licenses_count.empty())
-            g_licenses_manager->FreeLicense(job->licenses_count);
+            g_license_manager->FreeLicense(job->licenses_count);
           LockGuard indexes_guard(&m_job_indexes_mtx_);
           for (const CranedId& craned_id : job->CranedIds()) {
             m_node_to_jobs_map_[craned_id].erase(job->JobId());
@@ -4484,7 +4483,7 @@ void JobScheduler::CleanJobStatusChangeQueueCb_() {
         g_meta_container->FreeResourceFromResv(job->reservation, job->JobId());
       g_account_meta_container->FreeQosResource(*job);
       if (!job->licenses_count.empty())
-        g_licenses_manager->FreeLicense(job->licenses_count);
+        g_license_manager->FreeLicense(job->licenses_count);
 
       if (!g_config.JobLifecycleHook.CranectldEpilogs.empty()) {
         g_thread_pool->detach_task(
@@ -5551,7 +5550,7 @@ void SchedulerAlgo::NodeSelect(
                                           g_config.ScheduledBatchSize,
                                           job_ptr_vec);
 
-  g_licenses_manager->CheckLicenseCountSufficient(&job_ptr_vec);
+  g_license_manager->CheckLicenseCountSufficient(&job_ptr_vec);
 
   // Schedule pending jobs
   // TODO: do it in parallel
@@ -5899,7 +5898,7 @@ CraneExpected<void> JobScheduler::AcquireJobAttributes(JobInCtld* job) {
   }
 
   if (!job->JobToCtld().licenses_count().empty()) {
-    auto check_licenses_result = g_licenses_manager->CheckLicensesLegal(
+    auto check_licenses_result = g_license_manager->CheckLicensesLegal(
         job->JobToCtld().licenses_count(), job->JobToCtld().is_licenses_or());
     if (!check_licenses_result) {
       CRANE_ERROR("Failed to call CheckLicensesLegal: {}",
